backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==0.19.2
bcrypt==3.2.2
PyJWT==2.3.0
gunicorn==20.1.0
